# backend/public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sukify Documentation",
        "description": "Sukify Swagger OpenApi description",
        "contact": {
            "url": "https://github.com/Daizygod/sukify"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Demo API Server"
        },
        {
            "url": "https://sukify.ru",
            "description": "Deploy API Server"
        }
    ],
    "paths": {
...
        "/api/getaudio/{folder}/{filename}/{ext}": {
            "get": {
                "tags": [
                    "tracks"
                ],
                "summary": "Listen track",
                "description": "Listen track by file path",
                "operationId": "351baa601dce48f22eaa69556973e5cd",
                "parameters": [
                    {
                        "name": "folder",
                        "in": "path",
                        "description": "Folder of audio file",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filename",
                        "in": "path",
                        "description": "Filename of audio file",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ext",
                        "in": "path",
                        "description": "extension of audio file",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "audio/mpeg": {}
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Пошел нахуй"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    ]
}